+++ qiusheng-core/src/main/java/org/pockettech/qiusheng/impl/StoreListServiceImpl.java
@@ -1,6 +1,8 @@
 package org.pockettech.qiusheng.impl;
 
 import lombok.extern.slf4j.Slf4j;
+import org.pockettech.qiusheng.dao.ChartDao;
+import org.pockettech.qiusheng.entity.data.Chart;
 import org.pockettech.qiusheng.entity.data.Song;
 import org.pockettech.qiusheng.api.StoreListService;
 import org.pockettech.qiusheng.dao.SongDao;
@@ -8,12 +10,16 @@ import org.pockettech.qiusheng.entity.result.ListResult;
 import org.pockettech.qiusheng.entity.result.SongResult;
 import org.pockettech.qiusheng.entity.builder.StoreConditionalFilterBuilder;
 import org.pockettech.qiusheng.entity.filter.StoreConditionalFilter;
+import org.pockettech.qiusheng.entity.tools.ChartFileHandler;
 import org.pockettech.qiusheng.entity.tools.SongFilterHandle;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
+import java.io.File;
+import java.io.FileNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 //这里是返回歌曲列表的服务,以及推荐列表的服务
 @RestController
@@ -21,6 +27,8 @@ import java.util.List;
 public class StoreListServiceImpl implements StoreListService {
     @Resource
     SongDao songDao;
+    @Resource
+    ChartDao chartDao;
 
     StoreConditionalFilterBuilder builder = new StoreConditionalFilterBuilder();
 //  歌曲列表
@@ -89,17 +97,33 @@ public class StoreListServiceImpl implements StoreListService {
 
     @PostMapping("/admin/store/deleteSong")
     @ResponseBody
-    public int deleteSong(@RequestParam int sid) {
-        //TODO:添加删除歌曲以下所有铺面文件的选项与逻辑
+    public int deleteSong(@RequestParam int sid) throws FileNotFoundException {
+        File songFolder = new File(ChartFileHandler.getLocalFilePath() + "_song_" + String.valueOf(sid));
 
+        for (File f : Objects.requireNonNull(songFolder.listFiles()))
+            if (!f.delete())
+                return 1;
+
+        if (!songFolder.delete())
+            return 1;
+
+        //TODO:添加删除歌曲以下所有铺面文件的选项与逻辑
+        List<Chart> charts =chartDao.findSomeChartBySid(sid, 0, 50);
+        for (Chart chart: charts) {
+            if (chartDao.deleteChart(chart.getCid()) == 0) {
+                log.info("删除谱面(cid: " + chart.getCid() + ")时出错，可能原因为数据库中并未存在该谱面");
+                return 1;
+            } else
+                log.info("成功删除谱面(cid: " + chart.getCid() + ")");
+        }
 
         int deleteCode = songDao.deleteSong(sid);
 
         if (deleteCode == 0) {
-            log.info("删除谱面(cid: " + sid + ")时出错，可能原因为数据库中并未存在该谱面");
+            log.info("删除歌曲(sid: " + sid + ")时出错，可能原因为数据库中并未存在该歌曲");
             return 1;
         } else
-            log.info("成功删除谱面(cid: " + sid + ")");
+            log.info("成功删除歌曲(sid: " + sid + ")以及其下所有谱面文件和信息");
 
         return 0;
     }
